+++ src/index.js
@@ -1,10 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
+import { NewUserMandatory } from "./components";
 // import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
   <React.StrictMode>
+    <NewUserMandatory />
     <App />
   </React.StrictMode>,
   document.getElementById("app")
